+++ CRM/src/main/java/com/crm/model/Produto_.java
@@ -1,5 +1,6 @@
 package com.crm.model;
 
+import java.math.BigDecimal;
 import javax.annotation.Generated;
 import javax.persistence.metamodel.SingularAttribute;
 import javax.persistence.metamodel.StaticMetamodel;
@@ -9,7 +10,7 @@ import javax.persistence.metamodel.StaticMetamodel;
 public abstract class Produto_ {
 
 	public static volatile SingularAttribute<Produto, Categoria> categoria;
-	public static volatile SingularAttribute<Produto, Double> valor;
+	public static volatile SingularAttribute<Produto, BigDecimal> valor;
 	public static volatile SingularAttribute<Produto, String> nome;
 	public static volatile SingularAttribute<Produto, Long> id;
 	public static volatile SingularAttribute<Produto, Integer> quantidade;
